requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
click==6.7
colorama==0.4.1
coverage==4.5.3
crypto==1.4.1
cycler==0.10.0
django==2.2a1
Flask==1.0.2
gunicorn==19.9.0
idna==2.8
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
lxml==4.2.1
MarkupSafe==1.0
matplotlib==3.0.3
more-itertools==7.0.0
mpmath==1.1.0
mysql-connector==2.2.9
mysql-connector-python==8.0.13
Naked==0.1.31
numpy==1.16.2
pandas==0.24.2
pluggy==0.9.0
protobuf==3.6.1
py==1.8.0
pyparsing==2.4.0
pytest==4.4.1
python-dateutil==2.8.0
python-docx==0.8.6
pytz==2018.9
PyYAML==5.1
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
shellescape==3.4.1
six==1.12.0
sklearn==0.0
sqlparse==0.2.4
sympy==1.3
urllib3==1.24.1
Werkzeug==0.14.1
